+++ src/doc/index.ts
@@ -1,5 +1,12 @@
+import * as assert from 'assert';
+import chalk from 'chalk';
+import { copyFileSync, existsSync } from 'fs';
+import { join } from 'path';
+import ghPages from 'gh-pages';
 import * as docz from './docz';
 import * as storybook from './storybook';
+
+const DOC_PATH = '.doc';
 
 export interface DocProps {
   cwd: string;
@@ -7,6 +14,7 @@ export interface DocProps {
   params: string[];
   args: any;
   userConfig: any;
+  DOC_PATH?: typeof DOC_PATH;
 }
 
 export interface DeployProps {
@@ -15,15 +23,31 @@ export interface DeployProps {
 }
 
 export function devOrBuild(option: DocProps) {
-  if ((option || {})['--storybook']) {
-    return storybook.devOrBuild(option);
+  const mergedOption: DocProps = {
+    ...option,
+    DOC_PATH,
+  };
+
+  if ((option.args || {}).storybook) {
+    return storybook.devOrBuild(mergedOption);
   }
-  return docz.devOrBuild(option);
+  return docz.devOrBuild(mergedOption);
 }
 
-export function deploy(option: DeployProps) {
-  if ((option || {})['--storybook']) {
-    return storybook.deploy(option);
-  }
-  return docz.deploy(option);
+export function deploy({ cwd, args }) {
+  return new Promise((resolve, reject) => {
+    const distDir = join(cwd, DOC_PATH);
+
+    assert.ok(existsSync(distDir), `Please run ${chalk.green(`umi-lib doc build`)} first`);
+
+    copyFileSync(join(distDir, 'index.html'), join(distDir, '404.html'));
+
+    ghPages.publish(distDir, args, err => {
+      if (err) {
+        reject(new Error(`Doc deploy failed. ${err.message}`));
+      } else {
+        resolve();
+      }
+    });
+  });
 }
