+++ src/doc/storybook.ts
@@ -1,17 +1,14 @@
-import ghPages from 'gh-pages';
-import { join } from 'path';
 import { DocProps } from '.';
 import storybook from '@storybook/react/standalone';
 import generator from './storybook-generator';
 
-export function devOrBuild({ cwd, cmd }: Partial<DocProps>) {
+export function devOrBuild({ cwd, cmd, DOC_PATH }: Partial<DocProps>) {
   const { storybookPath } = generator(cwd);
 
   if (cmd === 'build') {
     return storybook({
       mode: 'static',
-      // 相对路径，storybook 会自动拼接 cmd 所在的位置
-      outputDir: './build',
+      outputDir: DOC_PATH,
       configDir: storybookPath,
     });
   } else {
@@ -22,19 +19,4 @@ export function devOrBuild({ cwd, cmd }: Partial<DocProps>) {
       configDir: storybookPath,
     });
   }
-}
-
-export async function deploy({ cwd, args }) {
-  const distDir = join(cwd, 'build');
-  await devOrBuild({ cwd, cmd: 'build' });
-
-  return new Promise((resolve, reject) => {
-    ghPages.publish(distDir, args, err => {
-      if (err) {
-        reject(new Error(`Doc deploy failed. ${err.message}`));
-      } else {
-        resolve();
-      }
-    });
-  });
 }
